Move hdfs functions to Impala backend
