+++ .github/workflows/main.yml
@@ -26,7 +26,7 @@ jobs:
       shell: bash
 
     - name: run tests
-      run: PYTEST_BACKENDS=$BACKENDS PYTEST_EXPRESSION="not hdfs" ./ci/run_tests.sh
+      run: PYTEST_BACKENDS=$BACKENDS ./ci/run_tests.sh
       shell: bash
 
   Tests_sql:
@@ -64,7 +64,7 @@ jobs:
       run: ./ci/setup_env.sh "${{ matrix.python_version }}" "$BACKENDS"
 
     - name: run tests
-      run: PYTEST_BACKENDS=$BACKENDS PYTEST_EXPRESSION="not udf and not hdfs" ./ci/run_tests.sh
+      run: PYTEST_BACKENDS=$BACKENDS PYTEST_EXPRESSION="not udf" ./ci/run_tests.sh
 
   Tests_impala_clickhouse:
     name: Tests Impala / Clickhouse
@@ -142,7 +142,7 @@ jobs:
       run: ./ci/setup_env.sh "${{ matrix.python_version }}" "$BACKENDS"
 
     - name: run tests
-      run: PYTEST_BACKENDS=$BACKENDS PYTEST_EXPRESSION="not hdfs" ./ci/run_tests.sh
+      run: PYTEST_BACKENDS=$BACKENDS ./ci/run_tests.sh
 
   Tests_spark:
     name: Tests PySpark / Spark
@@ -161,7 +161,7 @@ jobs:
       run: ./ci/setup_env.sh "${{ matrix.python_version }}" "$BACKENDS"
 
     - name: run tests
-      run: PYTEST_BACKENDS=$BACKENDS PYTEST_EXPRESSION="not hdfs" ./ci/run_tests.sh
+      run: PYTEST_BACKENDS=$BACKENDS ./ci/run_tests.sh
 
 
   Lint_and_benchmarks:
