+++ ibis/backends/impala/client.py
@@ -37,7 +37,7 @@ from ibis.backends.base_sql.ddl import (
 from ibis.backends.base_sqlalchemy.compiler import DDL, DML
 from ibis.client import Database, DatabaseEntity, Query, SQLClient
 from ibis.config import options
-from ibis.filesystems import HDFS, WebHDFS
+from .hdfs import HDFS, WebHDFS
 from ibis.util import log
 
 from . import ddl, udf
