+++ ibis/backends/impala/hdfs.py
